bicep/app-svc-python/requirements.txt
Flask==2.0.1
gevent==21.1.2
gunicorn==20.1.0
psycopg2-binary